requirements.txt
Flask>=2.3
python-docx>=0.8
